fix(review): finalize grafana and replay repository contracts
